fix(scheduler): remove tabindex from header element
