+++ packages/devextreme/js/__internal/scheduler/header/header.ts
@@ -4,6 +4,7 @@ import registerComponent from '@js/core/component_registrator';
 import devices from '@js/core/devices';
 import errors from '@js/core/errors';
 import $ from '@js/core/renderer';
+import { noop } from '@js/core/utils/common';
 import { getPathParts } from '@js/core/utils/data';
 import dateUtils from '@js/core/utils/date';
 import { extend } from '@js/core/utils/extend';
@@ -145,6 +146,8 @@ export class SchedulerHeader extends Widget<HeaderOptions> {
     this.renderToolbar();
     this._toggleVisibility();
   }
+
+  _renderFocusTarget(): void { return noop(); }
 
   private renderToolbar(): void {
     const config = this.createToolbarConfig();
